:lipstick: Componentize /index
