+++ jest.config.js
@@ -17,6 +17,7 @@ module.exports = {
     '^test(.*)$': resolve(__dirname, './test/$1'),
     '^pages(.*)$': resolve(__dirname, './pages/$1'),
     '^generated(.*)$': resolve(__dirname, './generated/$1'),
+    '^features(.*)$': resolve(__dirname, './features/$1'),
     '^mocks(.*)$': resolve(__dirname, './mocks/$1'),
     '\\.css$': resolve(__dirname, './test/tailwind-base-mock.js'),
   },
